Remove dependency on test/integration from kubemark
